Delivery Module Order Fragment  - 17 Auguest 2022 By Jilendra Singh

+++ app/src/main/java/naya/ganj/app/deliverymodule/model/DeliveryOrdersModel.kt
@@ -25,7 +25,7 @@ data class DeliveryOrdersModel(
         @SerializedName("paymentStatus")
         val paymentStatus: String,
         @SerializedName("totalAmount")
-        val totalAmount: Int,
+        val totalAmount: Double,
         @SerializedName("totalItems")
         val totalItems: Int
     ) {
